backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.1.0
boto3==1.29.7
celery==5.3.4
redis==5.0.1
django-filter==23.5
drf-spectacular==0.26.5
drf-spectacular[sidecar]==0.26.5
gunicorn==21.2.0
whitenoise==6.6.0
pytz==2023.3
qrcode==7.4.2
